+++ blue_geo/help/datacube/__init__.py
@@ -1,0 +1,13 @@
+from blue_geo.help.datacube.crop import help_crop
+from blue_geo.help.datacube.generate import help_generate
+from blue_geo.help.datacube.get import help_get
+from blue_geo.help.datacube.ingest import help_ingest, ingest_options, scope_details
+from blue_geo.help.datacube.list import help_list
+
+help_functions = {
+    "crop": help_crop,
+    "generate": help_generate,
+    "get": help_get,
+    "ingest": help_ingest,
+    "list": help_list,
+}
